arch: split providers into Geo & Territory
