+++ api/services/territory/src/interfaces/TerritoryRepositoryProviderInterface.ts
@@ -4,22 +4,14 @@ import {
   ResultInterface as CreateResultInterface,
 } from '../shared/territory/create.contract';
 import { ParamsInterface as PatchParamsInterface } from '../shared/territory/update.contract';
-import { ParamsInterface as DropdownParamsInterface } from '../shared/territory/listGeo.contract';
 import {
   TerritoryQueryInterface,
   SortEnum,
   ProjectionFieldsEnum,
   TerritoryListFilter,
 } from '../shared/territory/common/interfaces/TerritoryQueryInterface';
-import { TerritoryParentChildrenInterface } from '../shared/territory/common/interfaces/TerritoryChildrenInterface';
 import { ContactsInterface } from '../shared/common/interfaces/ContactsInterface';
 import { TerritoryLevelEnum } from '../shared/territory/common/interfaces/TerritoryInterface';
-import { TerritoryDropdownInterface } from '../shared/territory/common/interfaces/TerritoryDropdownInterface';
-
-import {
-  ParamsInterface as FindByInseeParamsInterface,
-  ResultInterface as FindByInseeResultInterface,
-} from '../shared/territory/findByInsees.contract';
 
 export interface TerritoryRepositoryProviderInterface {
   find(
@@ -29,7 +21,6 @@ export interface TerritoryRepositoryProviderInterface {
     pagination?: TerritoryListFilter,
   ): Promise<TerritoryDbMetaInterface>;
 
-  dropdown(params: DropdownParamsInterface): Promise<TerritoryDropdownInterface[]>;
   all(
     search?: string,
     levels?: TerritoryLevelEnum[],
@@ -43,8 +34,6 @@ export interface TerritoryRepositoryProviderInterface {
   update(data: PatchParamsInterface): Promise<TerritoryDbMetaInterface>;
   patchContacts(id: number, contacts: ContactsInterface): Promise<TerritoryDbMetaInterface>;
   patch(id: number, patch: { [k: string]: any }): Promise<TerritoryDbMetaInterface>;
-  findByInsees(params: FindByInseeParamsInterface): Promise<FindByInseeResultInterface>;
-  getDirectRelation(id: number | number[]): Promise<TerritoryParentChildrenInterface[]>;
 }
 
 export abstract class TerritoryRepositoryProviderInterfaceResolver implements TerritoryRepositoryProviderInterface {
@@ -54,14 +43,6 @@ export abstract class TerritoryRepositoryProviderInterfaceResolver implements Te
     projection: ProjectionFieldsEnum,
     pagination?: TerritoryListFilter,
   ): Promise<TerritoryDbMetaInterface> {
-    throw new Error();
-  }
-
-  async getDirectRelation(id: number | number[]): Promise<TerritoryParentChildrenInterface[]> {
-    throw new Error();
-  }
-
-  async dropdown(params: DropdownParamsInterface): Promise<TerritoryDropdownInterface[]> {
     throw new Error();
   }
 
@@ -93,10 +74,6 @@ export abstract class TerritoryRepositoryProviderInterfaceResolver implements Te
   }
 
   async patch(id: number, patch: { [k: string]: any }): Promise<TerritoryDbMetaInterface> {
-    throw new Error();
-  }
-
-  async findByInsees(params: FindByInseeParamsInterface): Promise<FindByInseeResultInterface> {
     throw new Error();
   }
 }
